Use pkg-config for Linux build.
pkg-config is a convenient way to have the right dependencies, no matter
what the distribution is and where the distributors put the headers/how
they named the library.

This fixes building dokidoki-support with Lua 5.1 on Debian Jessie.

Build command lines for MingW and MaxOSX have not been altered as I am
unaware of how easy it is to setup pkg-config on those platforms.